Check if steam vent is blocked when handling `lastRecipe`
Previously, venting is only checked when finding new recipe (line 245) by
calling `this.machine.modifyRecipe()`.
This means that when more than one item is processed, the `lastRecipe`
shortcuts will NOT check if venting is actually done, thus allowing
arbitrary number of items to be processed as long as the recipe never
changed (i.e. some of the same item is still in the input slot).