Implement deprecated getFoodProperties() method (Fixes Botania NPE)
Ideally, mods would use Forge's IForgeItem#getFoodProperties(ItemStack, LivingEntity) method.
However, this can be awkward for cross-loader mods, so we should provide a best-effort implementation
for Item#getFoodProperties() as well.

(Botania v446 and lower crashes without this; newer versions add a workaround, but the workaround
just assigns a nutrition of 0.)